requirements.txt
aiolimiter==1.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asyncio==3.4.3
cachetools==5.5.0
certifi==2024.6.2
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.0
coverage==5.5
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
eyed3==0.9.7
filetype==1.2.0
fonttools==4.54.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.7
jiter==0.8.2
kiwisolver==1.4.7
matplotlib==3.9.2
mcrcon==0.7.0
mutagen==1.47.0
nest-asyncio==1.6.0
Nuitka==2.5.9
numpy==2.1.1
openai==1.61.1
opencv-python==4.10.0.84
ordered-set==4.1.0
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytube==15.0.0
pytz==2024.2
pywin32-ctypes==0.2.2
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.3.0
zstandard==0.23.0
